Task_eight
Define a Bartender class that has a createDrink method that creates a drink with any number of user-specified ingredients. As a result, the method should return an object of the Drink type with information about the ingredients of a given drink. Each ingredient in the Drink class should be represented by the Ingredient class, which stores the name and amount of the ingredient. Load the data from the user using the application.

The Bartender class should also have a printDrink method that takes an object of the Drink type as a parameter and displays information about it.