Task_nine
In the program, create a class that represents a single song. Each song should have a name and duration in seconds.

In a separate class, such as Playlist, define a method that takes any number of songs (as separate arguments) and returns the total playing time in seconds. So it should be possible to call the method in any of the following ways: