Task_ten
Write a program that will simulate the role of a ticket controller. Tickets are allocated in electronic form and have a limited validity period.

Define the classes:

Ticket - represents the ticket, stores information about the time of validation of the ticket and its validity time (expressed in minutes). Since tickets are sold only in electronic form, the ticket is immediately "validated" after purchase,
Passenger representing the passenger. It stores information about the name and surname. They can also be assigned a Ticket object, which means that the passenger has purchased a ticket.
TicketMachine - Represents a ticket machine. It has only one buyTicket method, which takes as arguments the passenger and the time for which the passenger buys the ticket, and its task is to create a ticket and assign it to the passenger's property transferred.
In the TicketController class, which should be the starting class of the program, create an array of passengers and assign tickets to some of them (some of which should be "expired"). Define a separate method to find out which passengers have a valid ticket. As a result, the method should return an array of passengers who do not have a valid ticket.