Task_airConditioners
The building consists of 4 rooms, and each room can have an independent air conditioner of one of two types:

BasicAirConditioner - lowers the room temperature by 1 degree per minute with a capacity of 1 cubic meter per minute. For example, if a room has a volume of 10 cubic meters, the room temperature will be lowered by 0.1 degrees per minute.
ProAirConditioner - lowers the room temperature by 2 degrees per minute with a capacity of 1 cubic meter per minute. For example, if the room has a volume of 10 cubic meters, the room temperature will be lowered by 0.2 degrees per minute.
For the purposes of the task, we simulate that 1 minute corresponds to 1 second.

In each room you can set the air conditioning on or off and the temperature to which the temperature is to be reduced.

The program should display the temperature in each of the four rooms every second. Every second the temperature in the room should be lowered by an appropriate value depending on the air conditioner used. Assume any starting temperature. After lowering the temperature to the set value, it should be kept constant.